Refactor: Resolve compilation errors and refactor package imports
This commit resolves a large number of compilation errors that were causing the build to fail. The fixes primarily involve correcting package imports and resolving symbol references across multiple modules.

Key changes include:
- Corrected the import path for `AuraAIService` in `TrinityCoordinatorService` and `GenesisAgent`.
- Updated the import path for `NeuralWhisper` in `ConferenceRoomViewModel`.
- Resolved numerous `Unresolved reference` errors throughout the codebase by fixing incorrect or missing imports for classes like `AgentResponse`, `GenKitUiState`, `EnhancedInteractionData`, and various UI components.
- Addressed errors related to unimplemented interface members in several `Agent` classes.
- Fixed type mismatches, incorrect function overrides, and invalid parameter arguments.
