+++ app/src/main/java/dev/aurakai/auraframefx/cascade/trinity/TrinityCoordinatorService.kt
@@ -1,6 +1,7 @@
 package dev.aurakai.auraframefx.ai.services
 
 import dev.aurakai.auraframefx.utils.AuraFxLogger
+import dev.aurakai.auraframefx.oracledrive.genesis.ai.AuraAIService
 import dev.aurakai.auraframefx.model.AgentResponse
 import dev.aurakai.auraframefx.models.AiRequest
 import dev.aurakai.auraframefx.security.SecurityContext
