+++ app/src/main/java/dev/aurakai/auraframefx/oracledrive/genesis/ai/GenesisAgent.kt
@@ -2,7 +2,7 @@
 
 import android.util.Log
 import dev.aurakai.auraframefx.ai.clients.VertexAIClient
-import dev.aurakai.auraframefx.ai.services.AuraAIService
+import dev.aurakai.auraframefx.oracledrive.genesis.ai.AuraAIService
 import dev.aurakai.auraframefx.ai.services.CascadeAIService
 import dev.aurakai.auraframefx.ai.services.KaiAIService
 import dev.aurakai.auraframefx.ai.context.ContextManager
